LayoutTests/platform/mac/accessibility/aria-menu-role-descriptions.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<div role="menu" tabindex="0" id="menu">
<div role="menuitem" tabindex="0" id="menuitem">menu item</div>
</div>


<p id="description"></p>
<div id="console"></div>

<script>

    description("This makes sure that AXMenu and AXMenuItem have role descriptions.");

    if (window.accessibilityController) {

          var menu = document.getElementById("menu");
          menu.focus();

          menu = accessibilityController.focusedElement;
          var menuRole = menu.stringAttributeValue("AXRoleDescription");
          shouldBe("menuRole.length > 0", "true");

          var menuItem = menu.childAtIndex(0);
          var menuItemRole = menuItem.stringAttributeValue("AXRoleDescription");
          shouldBe("menuItemRole.length > 0", "true");
    }

</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
