safari-534.53-branch/LayoutTests/platform/mac/accessibility/aria-menu-role-descriptions-expected.txt

menu item
This makes sure that AXMenu and AXMenuItem have role descriptions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS menuRole.length > 0 is true
PASS menuItemRole.length > 0 is true
PASS successfullyParsed is true

TEST COMPLETE
